mdns: Split DNS parsing and CaseInsensitiveDict into separate modules
